refactor: apply-domain-driven-architecture
