--- src/main/java/com/bluewhaletech/Ourry/infrastructure/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.bluewhaletech.Ourry.infrastructure.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
