+++ src/main/java/com/bluewhaletech/Ourry/infrastructure/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.bluewhaletech.Ourry.infrastructure.config;
+package com.bluewhaletech.Ourry.infrastructure.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
