+++ src/main/java/com/bluewhaletech/Ourry/infrastructure/security/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.bluewhaletech.Ourry.infrastructure.properties;
+package com.bluewhaletech.Ourry.infrastructure.security;
 
 import lombok.Getter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
