+++ src/main/java/com/bluewhaletech/Ourry/presentation/GlobalExceptionHandler.java
@@ -1,5 +1,8 @@
-package com.bluewhaletech.Ourry.common.exception;
+package com.bluewhaletech.Ourry.presentation;
 
+import com.bluewhaletech.Ourry.common.exception.BusinessException;
+import com.bluewhaletech.Ourry.common.exception.ErrorCode;
+import com.bluewhaletech.Ourry.common.exception.ErrorResponse;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.UnsupportedJwtException;
