+++ src/main/resources/application.yml
@@ -1,14 +1,18 @@
 spring:
   datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
     url: URL
     username: USERNAME
     password: PASSWORD
+    driver-class-name: org.postgresql.Driver
   jpa:
     database: mysql
     show-sql: true
     generate-ddl: true
     hibernate:
       ddl-auto: create
+    properties:
+      show_sql: true
+      format_sql: true
+      dialect: org.hibernate.dialect.PostgreSQLDialect
   profiles:
     active: local
