+++ src/test/java/com/bluewhaletech/Ourry/OurryApplicationTests.java
@@ -4,10 +4,10 @@ import com.bluewhaletech.Ourry.domain.article.repository.QuestionRepository;
 import com.bluewhaletech.Ourry.domain.member.repository.MemberRepository;
 import com.bluewhaletech.Ourry.domain.auth.repository.RedisJwtRepository;
 import com.bluewhaletech.Ourry.domain.poll.repository.*;
-import com.bluewhaletech.Ourry.infrastructure.jwt.JwtProvider;
+import com.bluewhaletech.Ourry.infrastructure.security.JwtProvider;
 import com.bluewhaletech.Ourry.application.member.MemberService;
-import com.bluewhaletech.Ourry.infrastructure.util.RedisBlackListManagement;
-import com.bluewhaletech.Ourry.infrastructure.util.RedisEmailAuthentication;
+import com.bluewhaletech.Ourry.infrastructure.redis.RedisBlackListManagement;
+import com.bluewhaletech.Ourry.infrastructure.redis.RedisEmailAuthentication;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
